Fixes #2, rm code splits for server side rendering
Using the `ONSERVER` constant defined in both the
server and client webpack configs, we can determine
if the code is targeted for client or server execution.

In this case, we need to remove code splitting on the
server side so that server side rendering works correctly.
Ie. `System.import` and `require.ensure` are not resolved
in the server environment.