Transition from react-router 3 to 4 (breaking)
This is a major revision of react-router. I agree, the churn is
high and can be annoying, but after upgrading this demo to work
with react-router 4, I will say I'm a fan. It feels much cleaner
and more es6y ;).

So here it is, the client side with code splitting is working.

SERVER SIDE NOT WORKING YET. They say it can't be done, but I
hope I can get it working!

"Godspeed those who attempt the server-rendered, code-split apps."
- https://reacttraining.com/react-router/web/guides/code-splitting